containers/read-source-data/requirements.txt
azure-identity
azure-mgmt-datafactory
azure-storage-blob
fastapi
uvicorn
httpx
phdi